refactor(eslint-plugin): 🛠 factorize rule tester
